+++ include/LZ77/LZ77BinCoder.hpp
@@ -35,7 +35,7 @@ public:
         }
         else {
             if(m_bitout.writeFrom(1, 1) != 1 ||
-                m_bitout.writeFrom(std::get<size_t>(p_value.value), factor_bit_size) != factor_bit_size ||
+                m_bitout.writeFrom(std::get<u_int32_t>(p_value.value), factor_bit_size) != factor_bit_size ||
                 m_bitout.writeFrom(p_value.length, factor_bit_size) != factor_bit_size) {
                  return 0;
             }
@@ -84,7 +84,7 @@ public:
             m_bit_counter += 9;
         }
         else {
-            size_t offset;
+            u_int32_t offset;
             if(m_bitin.readInto(offset,factor_bit_size) !=factor_bit_size ||
                 m_bitin.readInto(p_value.length, factor_bit_size) != factor_bit_size) {
                 return 0;
